Refactor reading of custom mappings
